Update years.json
Adding 2023 ERAs C and D prompt. I would like to understand the behavior of the code. Since this is not a ReRECO and we are dealing only with prompt, the code would need to append all 4 versions of the promptReco to the v1 of RAW. This is yet to be seen.